requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests>=2.32.2,<3
gunicorn==21.2.0
Werkzeug==2.3.7
langchain==0.2.14
langchain-community==0.2.12
chromadb==0.5.3
docling==2.70.0
ibm-watsonx-ai==1.0.5
sentence-transformers==2.6.1
